requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
langchain==0.3.14
langchain-openai==0.2.14
langchain-core==0.3.28
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.11
chromadb==0.5.23
sentence-transformers==3.3.1
unstructured==0.16.14
unstructured[pdf]==0.16.14
pydantic==2.10.5
python-dotenv==1.0.1
aiofiles==24.1.0
tiktoken==0.8.0
